dailyRanking now are only associated with contents and not chunks, like it was before. This is sub-optimal in terms of accuracy, but it fixes a bug related to the fact that the number of chunks generated each day is not constant (unlike contents) and thus the previous method was broken.
Also fixed an error in the Debug build configuration.

Now there is a bug related to flows which are carried over the next round; at the beginning of a new round generateViewMap is called and the content of the userWatchMap is overwritten, so when the carried over flows are completed we make a mess. I need to handle these carried over flows differently, especially since they are supposed to be related to the last chunk of some content and so it should not be a big deal